Sonar: Dodgy - int division result cast to double or float
git-svn-id: svn://localhost/jahia/trunk@24056 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ core/src/main/java/org/jahia/services/cache/CacheHelper.java
@@ -269,7 +269,7 @@ public final class CacheHelper {
         StatisticsGateway stats = cache.getStatistics();
         info.setHitCount(stats.cacheHitCount());
         info.setMissCount(stats.cacheMissCount());
-        info.setHitRatio(info.getAccessCount() > 0 ? info.getHitCount() * 100 / info.getAccessCount() : 0);
+        info.setHitRatio(info.getAccessCount() > 0 ? info.getHitCount() * 100 / (double) info.getAccessCount() : 0);
 
         info.setSize(stats.getSize());
